requirements.txt
pyzmq
six
